+++ festival/src/main/java/com/example/festival/auth/repository/AuthRepository.java
@@ -3,7 +3,7 @@ package com.example.festival.auth.repository;
 import com.example.festival.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthRepository extends JpaRepository<User, Long> {
+public interface AuthRepository extends JpaRepository<User, Integer> {
     //User findByUsername(String username);
     User findByIdentify(String identify);
     boolean existsByIdentify(String identify);
