+++ festival/src/main/java/com/example/festival/comment/repository/CommentRepositoryInterface.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CommentRepositoryInterface extends CrudRepository<Comment, Long> {
-    public List<Comment> findAllByPartner_PartnerId(Long partnerId);
-    public int countCommentsByPartner_PartnerId(Long partnerId);
+public interface CommentRepositoryInterface extends CrudRepository<Comment, Integer> {
+    public List<Comment> findAllByPartner_PartnerId(Integer partnerId);
+    public int countCommentsByPartner_PartnerId(Integer partnerId);
 }
