+++ festival/src/main/java/com/example/festival/reply/dto/ReplyDto.java
@@ -10,10 +10,10 @@ import java.sql.Timestamp;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReplyDto {
-    private Long replyId;
+    private Integer replyId;
     private String content; //내용
     private String nickname; //작성자
     private String address; //작성자 주소
-    private Long commentId; //댓글
+    private Integer commentId; //댓글
     private Timestamp createdAt;
 }
